Combined individual core plugin tests into a single translation unit and header. Fixed minor issues with workspace testing.
