+++ tests/Makefile.am
@@ -18,8 +18,7 @@ pkginclude_HEADERS = system_tests.h \
                      workspace_tests.h \
                      plugin_tests.h \
                      fakePlugin.h \
-                     performance_measurement_plugin_tests.h \
-                     connector_plugin_tests.h
+                     core_plugins_tests.h
 
 rtxiTests_SOURCES = rtxiTests.cpp \
                     system_tests.cpp \
@@ -29,8 +28,7 @@ rtxiTests_SOURCES = rtxiTests.cpp \
                     settings_tests.cpp \
                     workspace_tests.cpp \
                     plugin_tests.cpp \
-                    performance_measurement_plugin_tests.cpp \
-                    connector_plugin_tests.cpp
+                    core_plugins_tests.cpp
 
 OBJS = $(top_srcdir)/src/rt.o \
        $(top_srcdir)/src/event.o \
@@ -45,7 +43,6 @@ OBJS = $(top_srcdir)/src/rt.o \
        $(top_srcdir)/src/workspace.o \
        $(top_srcdir)/src/main_window.o \
        $(top_srcdir)/src/default_gui_model.o \
-       $(top_srcdir)/plugins/performance_measurement/*.o \
        $(top_srcdir)/libs/math/runningstat.o \
        $(top_srcdir)/src/moc_*.o
