--- tests/connector_plugin_tests.cpp
@@ -1,30 +1,0 @@
-/*
- 	 The Real-Time eXperiment Interface (RTXI)
-	 Copyright (C) 2011 Georgia Institute of Technology, University of Utah, Weill Cornell Medical College
-
-	 This program is free software: you can redistribute it and/or modify
-	 it under the terms of the GNU General Public License as published by
-	 the Free Software Foundation, either version 3 of the License, or
-	 (at your option) any later version.
-
-	 This program is distributed in the hope that it will be useful,
-	 but WITHOUT ANY WARRANTY; without even the implied warranty of
-	 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	 GNU General Public License for more details.
-
-	 You should have received a copy of the GNU General Public License
-	 along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
- */
-
-#include <connector_plugin_tests.h>
-
-TEST_F(ConnectorPluginTests, loadPlugin)
-{
-    // TODO: decouple plugin loading from QT gui framework for unit testing
-}
-
-TEST_F(ConnectorPluginTests, getInstance)
-{
-    // TODO: decouple getInstance function from QT gui framework for unit testing
-}
